Rename app from app_api to api
