+++ api/models/__init__.py
@@ -19,7 +19,7 @@ Base = declarative_base(cls=BaseMixin)
 
 
 # all models, for which tables should be created, must be listed here:
-from app_api.models import document  # noqa
-from app_api.models import waypoint  # noqa
-from app_api.models import route  # noqa
-from app_api.models import document_history  # noqa
+from api.models import document  # noqa
+from api.models import waypoint  # noqa
+from api.models import route  # noqa
+from api.models import document_history  # noqa
