+++ api/tests/views/test_waypoint.py
@@ -1,36 +1,36 @@
 import json
 
-from app_api.models.route import Route, RouteLocale
+from api.models.waypoint import Waypoint, WaypointLocale
 
 from .. import BaseTestCase
 
 
-class TestRouteRest(BaseTestCase):
+class TestWaypointRest(BaseTestCase):
 
     def setUp(self):  # noqa
         BaseTestCase.setUp(self)
         self._add_test_data()
 
     def test_get_collection(self):
-        response = self.app.get('/routes')
+        response = self.app.get('/waypoints')
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.content_type, 'application/json')
 
         body = json.loads(response.body)
         self.assertTrue(isinstance(body, list))
-        nb_routes = self.session.query(Route).count()
-        self.assertEqual(len(body), nb_routes)
+        nb_waypoints = self.session.query(Waypoint).count()
+        self.assertEqual(len(body), nb_waypoints)
 
     def test_get(self):
-        response = self.app.get('/routes/' + str(self.route.document_id))
+        response = self.app.get('/waypoints/' + str(self.waypoint.document_id))
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.content_type, 'application/json')
 
         body = json.loads(response.body)
         self.assertFalse('id' in body)
-        self.assertEqual(body.get('document_id'), self.route.document_id)
+        self.assertEqual(body.get('document_id'), self.waypoint.document_id)
         self.assertEqual(
-            body.get('activities'), self.route.activities)
+            body.get('waypoint_type'), self.waypoint.waypoint_type)
 
         locales = body.get('locales')
         self.assertEqual(len(locales), 1)
@@ -42,7 +42,7 @@ class TestRouteRest(BaseTestCase):
     def test_post_error(self):
         body = {}
         response = self.app.post(
-            '/routes', params=json.dumps(body), expect_errors=True)
+            '/waypoints', params=json.dumps(body), expect_errors=True)
         self.assertEqual(response.status_code, 400)
 
         body = json.loads(response.body)
@@ -50,19 +50,19 @@ class TestRouteRest(BaseTestCase):
         errors = body.get('errors')
         self.assertEqual(len(errors), 1)
         self.assertEqual(
-            errors[0].get('description'), 'activities is missing')
-        self.assertEqual(errors[0].get('name'), 'activities')
+            errors[0].get('description'), 'waypoint_type is missing')
+        self.assertEqual(errors[0].get('name'), 'waypoint_type')
 
     def test_post_missing_title(self):
         body = {
-            'activities': 'skitouring',
-            'height': 1200,
+            'waypoint_type': 'summit',
+            'elevation': 3200,
             'locales': [
                 {'culture': 'en'}
             ]
         }
         response = self.app.post(
-            '/routes', params=json.dumps(body), expect_errors=True,
+            '/waypoints', params=json.dumps(body), expect_errors=True,
             content_type='application/json')
         self.assertEqual(response.status_code, 400)
 
@@ -75,15 +75,15 @@ class TestRouteRest(BaseTestCase):
 
     def test_post_success(self):
         body = {
-            'activities': 'hiking',
-            'height': 750,
+            'waypoint_type': 'summit',
+            'elevation': 3779,
             'locales': [
-                {'culture': 'en', 'title': 'Some nice loop',
-                 'gear': 'shoes'}
+                {'culture': 'en', 'title': 'Mont Pourri',
+                 'pedestrian_access': 'y'}
             ]
         }
         response = self.app.post(
-            '/routes', params=json.dumps(body),
+            '/waypoints', params=json.dumps(body),
             content_type='application/json')
         self.assertEqual(response.status_code, 200)
 
@@ -92,8 +92,8 @@ class TestRouteRest(BaseTestCase):
         self.assertIsNotNone(document_id)
 
         # check that the version was created correctly
-        route = self.session.query(Route).get(document_id)
-        versions = route.versions
+        waypoint = self.session.query(Waypoint).get(document_id)
+        versions = waypoint.versions
         self.assertEqual(len(versions), 1)
         version = versions[0]
 
@@ -106,26 +106,26 @@ class TestRouteRest(BaseTestCase):
         self.assertEqual(meta_data.comment, 'creation')
         self.assertIsNotNone(meta_data.written_at)
 
-        archive_route = version.document_archive
-        self.assertEqual(archive_route.document_id, document_id)
-        self.assertEqual(archive_route.activities, 'hiking')
-        self.assertEqual(archive_route.height, 750)
+        archive_waypoint = version.document_archive
+        self.assertEqual(archive_waypoint.document_id, document_id)
+        self.assertEqual(archive_waypoint.waypoint_type, 'summit')
+        self.assertEqual(archive_waypoint.elevation, 3779)
 
         archive_locale = version.document_locales_archive
         self.assertEqual(archive_locale.document_id, document_id)
         self.assertEqual(archive_locale.culture, 'en')
-        self.assertEqual(archive_locale.title, 'Some nice loop')
-        self.assertEqual(archive_locale.gear, 'shoes')
+        self.assertEqual(archive_locale.title, 'Mont Pourri')
+        self.assertEqual(archive_locale.pedestrian_access, 'y')
 
     def _add_test_data(self):
-        self.route = Route(
-            activities='paragliding', height=2000)
+        self.waypoint = Waypoint(
+            waypoint_type='summit', elevation=2203)
 
-        self.locale_en = RouteLocale(
-            culture='en', title='Mont Blanc from the air', description='...',
-            gear='paraglider')
+        self.locale_en = WaypointLocale(
+            culture='en', title='Mont Granier', description='...',
+            pedestrian_access='yep')
 
-        self.route.locales.append(self.locale_en)
+        self.waypoint.locales.append(self.locale_en)
 
-        self.session.add(self.route)
+        self.session.add(self.waypoint)
         self.session.flush()
