+++ api/views/waypoint.py
@@ -1,9 +1,9 @@
 from cornice.resource import resource, view
 from sqlalchemy.orm import joinedload
 
-from app_api.models.waypoint import Waypoint, schema_waypoint
-from app_api.models import DBSession
-from app_api.views.document import DocumentRest
+from api.models.waypoint import Waypoint, schema_waypoint
+from api.models import DBSession
+from api.views.document import DocumentRest
 from . import validate_id, to_json_dict
